style: improve code-style
