+++ dialog.h
@@ -33,7 +33,7 @@ private:
     vector<string> cities_list;
 };
 
-void show_winnerwindow(Dialog* dialog_ptr, cities::game_result result);
+void show_endgamewindow(Dialog* dialog_ptr, cities::game_result result);
 void cities_game();
 
 #endif // DIALOG_H
