+++ naming_cities.txt
@@ -1,0 +1,273 @@
+abidjan
+abu dhabi
+abuja
+accra
+addis ababa
+adelaide
+alexandria
+algiers
+amman
+amsterdam
+ankara
+antananarivo
+antwerp
+asheville
+asuncion
+athens
+auckland
+baghdad
+baku
+bamako
+bangkok
+barcelona
+basel
+beijing
+beirut
+belfast
+belgrade
+bengaluru
+berlin
+bilbao
+bogota
+boston
+brasilia
+brasov
+bratislava
+bremen
+brisbane
+brno
+brussels
+bucharest
+budapest
+buenos aires
+buffalo
+bydgoszcz
+cagliari
+cairo
+calgary
+canberra
+cape town
+caracas
+casablanca
+chandigarh
+charlotte
+chattanooga
+chengdu
+chicago
+chisinau
+cluj-napoca
+colombo
+columbus
+conakry
+copenhagen
+coventry
+dakar
+dallas
+damascus
+dar es salaam
+darwin
+delhi
+denver
+detroit
+dhaka
+djibouti
+dodoma
+doha
+dubai
+dublin
+durban
+edinburgh
+essen
+faisalabad
+fukuoka
+gaborone
+gaziantep
+genoa
+giza
+gothenburg
+graz
+guangzhou
+guatemala city
+hamburg
+hamilton
+hanoi
+harare
+havana
+helsinki
+homs
+hong kong
+houston
+hyderabad
+ingolstadt
+irkutsk
+islamabad
+istanbul
+izmir
+jaipur
+jakarta
+jeddah
+jefferson city
+jinan
+johannesburg
+johor bahru
+juarez
+kampala
+kano
+karachi
+kathmandu
+kolkata
+krakow
+kuala lumpur
+kuwait city
+kyiv
+la paz
+lagos
+lahore
+leeds
+leipzig
+lille
+lima
+limoges
+lisbon
+liverpool
+ljubljana
+luxembourg
+lyon
+madrid
+malaga
+malmo
+managua
+manama
+mandalay
+manila
+maputo
+marrakech
+mashhad
+medan
+medellin
+melbourne
+mendoza
+mexico city
+miami
+milan
+minsk
+mogadishu
+monrovia
+montevideo
+montreal
+moscow
+mumbai
+munich
+nagoya
+nairobi
+nashik
+nashville
+nassau
+new delhi
+new york
+newcastle
+niamey
+nicosia
+nizhny novgorod
+norwich
+nottingham
+odessa
+okayama
+orlando
+osaka
+oslo
+ottawa
+ouagadougou
+palermo
+panama city
+paramaribo
+paris
+patna
+perth
+phnom penh
+phuket
+pittsburgh
+port louis
+port moresby
+poznan
+prague
+pretoria
+puebla
+pune
+pyongyang
+quanzhou
+quito
+rabat
+recife
+reykjavik
+riga
+rio de janeiro
+riyadh
+rome
+rotterdam
+salt lake city
+salzburg
+samarkand
+san jose
+san juan
+san salvador
+santiago
+santo domingo
+sao paulo
+sapporo
+sarajevo
+sekondi-takoradi
+sendai
+seoul
+seville
+sheffield
+singapore
+skopje
+sofia
+stockholm
+stuttgart
+surat
+sydney
+taichung
+taipei
+tallinn
+tangier
+tbilisi
+tegucigalpa
+tehran
+tel aviv
+thimphu
+tianjin
+tirana
+tokyo
+toluca
+toronto
+touba
+tunis
+turin
+ulaanbaatar
+ulsan
+urumqi
+vaduz
+valencia
+valletta
+vancouver
+victoria
+vienna
+vilnius
+vitoria-gasteiz
+volgograd
+warsaw
+wellington
+windhoek
+wroclaw
+wuhan
+xalapa
+xiamen
+yangon
+yaounde
+yerevan
+yokohama
+zagreb
+zanzibar
+zurich
