+++ resources/mainwindow.pro.user
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE QtCreatorProject>
-<!-- Written by QtCreator 13.0.2, 2024-07-19T21:02:22. -->
+<!-- Written by QtCreator 13.0.2, 2024-07-25T11:57:54. -->
 <qtcreator>
  <data>
   <variable>EnvironmentId</variable>
