--- ui/winnerwindow.ui
@@ -1,95 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>Winnerwindow</class>
- <widget class="QDialog" name="Winnerwindow">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>400</width>
-    <height>300</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Cities</string>
-  </property>
-  <property name="windowIcon">
-   <iconset resource="resources.qrc">
-    <normaloff>:/img/cities_icon.png</normaloff>:/img/cities_icon.png</iconset>
-  </property>
-  <property name="styleSheet">
-   <string notr="true">QDialog {
-	background-color: green;
-}</string>
-  </property>
-  <widget class="QPushButton" name="restart_button">
-   <property name="geometry">
-    <rect>
-     <x>50</x>
-     <y>250</y>
-     <width>83</width>
-     <height>29</height>
-    </rect>
-   </property>
-   <property name="styleSheet">
-    <string notr="true">QPushButton {
-	background-color: rgb(0, 85, 0);
-}
-
-QPushButton:hover {
-	background-color:  rgb(0, 102, 0);
-}
-
-QPushButton:pressed {
-	background-color:  rgb(0, 124, 0);
-}</string>
-   </property>
-   <property name="text">
-    <string>Restart</string>
-   </property>
-  </widget>
-  <widget class="QPushButton" name="exit_button">
-   <property name="geometry">
-    <rect>
-     <x>260</x>
-     <y>250</y>
-     <width>83</width>
-     <height>29</height>
-    </rect>
-   </property>
-   <property name="styleSheet">
-    <string notr="true">QPushButton {
-	background-color: rgb(0, 85, 0);
-}
-
-QPushButton:hover {
-	background-color:  rgb(0, 102, 0);
-}
-
-QPushButton:pressed {
-	background-color:  rgb(0, 124, 0);
-}</string>
-   </property>
-   <property name="text">
-    <string>Exit</string>
-   </property>
-  </widget>
-  <widget class="QLabel" name="game_result">
-   <property name="geometry">
-    <rect>
-     <x>110</x>
-     <y>70</y>
-     <width>171</width>
-     <height>61</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;&lt;br/&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
-   </property>
-  </widget>
- </widget>
- <resources>
-  <include location="resources.qrc"/>
- </resources>
- <connections/>
-</ui>
